Better organise datatable directives
