+++ src/app/directives/datatable/datatable.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from "@angular/core";
-import { DatatableSortKeyDirective } from "./sort-key.directive";
-import { DatatableDefaultsDirective } from "./defaults.directive";
-import { DatatablePaginationDirective } from "./pagination.directive";
-import { VirtualDatatablePaginationDirective } from "./virtualDatatablePagination/virtualDatatablePagination.directive";
+import { DatatableSortKeyDirective } from "./sort-key/sort-key.directive";
+import { DatatableDefaultsDirective } from "./defaults/defaults.directive";
+import { DatatablePaginationDirective } from "./pagination/pagination.directive";
+import { VirtualDatatablePaginationDirective } from "./virtualDatatablePagination/virtual-datatable-pagination.directive";
 
 const directives = [
   DatatableDefaultsDirective,
