+++ src/app/directives/datatable/sort-key/sort-key.directive.spec.ts
@@ -6,7 +6,7 @@ import {
 } from "@swimlane/ngx-datatable";
 import { getCallArgs } from "@test/helpers/general";
 import { DatatableSortKeyDirective } from "./sort-key.directive";
-import { DatatableSortEvent } from "./pagination.directive";
+import { DatatableSortEvent } from "../pagination/pagination.directive";
 
 describe("DatatableSortKeyDirective", () => {
   let spec: SpectatorDirective<DatatableComponent>;
